Extract file encoding issue helpers partial
